# [Perf] Time.iso8601 is 3x faster than Time.parse