--- a/gems/aws-sdk-core/lib/aws-sdk-core/ecs_credentials.rb
+++ b/gems/aws-sdk-core/lib/aws-sdk-core/ecs_credentials.rb
@@ -85,7 +85,7 @@ module Aws
           c['SecretAccessKey'],
           c['Token']
         )
-        @expiration = c['Expiration'] ? Time.parse(c['Expiration']) : nil
+        @expiration = c['Expiration'] ? Time.iso8601(c['Expiration']) : nil
       end
     end