@awood45 2018-06-28
Looks good, some administrivia: Can you confirm "By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license."